Singing blocks prototype
I have an object being created with a variable definition that can change the y_scale of the overall base sprite. Then when the oNote object is colliding with the oMusicBlock object I first set the pitch of the sound to be scaled by the instances y_scale variable, then the sound plays.